Create Decision-trees
Building a decision tree applied to historical data of patients, and their response to different medications. Afterwards, the trained decision tree can be used to predict the class of an unknown patient, or to find a proper drug for a new patient.

The algorithm is obtained from the Scikit-Learn library (Python).